move webview stuff into submodule for cleaner main class
